requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
watchdog
ultralytics
torch
opencv-python-headless
python-dotenv
